merge: accept objects instead of commits, rename functions
Prepare to implement squashing where it will be needed to merge in
series and the result is not a commit.

Refactoring, no functional changes.

Make main merging function accept Object instead of concret
Commit (supposed to be either Commit or a Tree) and peel it to tree
instread of taking the tree from Commits.

Commits are still main objects in the high level functions since
parents are used to find merge bases.

Rename the functions in more structural way merge_<what>_<result>,
just looks more complete.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>