cmp_commits: implement squashing of input commits
Make it possible to compare squashed commits.

Change interface to accept list of commits after the first one, which
is "others", squash them and compare the resulting trees.

Use the same logic for one commit as well, it will function the same
way as it did before.

Squashig works with cherry-picking all the commits in the list to the
first commits (by merging with parent as the base).

Use .and_then() iterator method to stop merging and propogate errors.

Since writing index returns not Tree object, but just Oid (to create a
commit based on than), have to use find_object() to continue
iterations, but return final Oid without amending or extra lookups.

Update README.md to be consistent with the code.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>