cmp_commit: implement autofetch
implement boolean --autofetch switch with which it reads message of
the commit under investigation and fetches upstream commit
ids (likely squashed) in form

`(cherry picked from <sha1>)`

or
`commit <sha1>`

(both at the beginning of the line)

The commits are added in order from the top of the message.

If more than one commit provided as arguments, they are squashed
before the automatically fetched ones.

If --autofetch is supplied, the head is not added to the squash list.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>